Fix missing "s" in url path to display examples
**Description**
The url path to display the examples in the repo was missing an "s". For example, the path `"./example/pokeapi_functional/"` would point to a 404 page because the actual path should have been `"./examples/pokeapi_functional/"`.

**Changes**
I have updated the url path to display the examples to include the missing "s". This change has been tested and works as expected.